requirements.txt
google-generativeai>=0.8.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14
pyttsx3>=2.90
